将连接预热的生命周期拆到独立文件
SessionStore.swift 因新增预热 API 超过 2000 行体积门禁。类里只保留存储，
begin/end 与 isEstablishingConnection 的判定移入 SessionStoreConnectionWarmUp。

Refs #403
